+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project uses [Semantic Versioning][sv].
 ## [Unreleased][new]
 
 ### Added
+- Add routines to load, normalize, and partition data from an H5 file into
+  training and test sets.
 
 ### Changed